advancing part 22

+++ src/comments/views.py
@@ -1,6 +1,6 @@
 from django.contrib import messages
 from django.contrib.contenttypes.models import ContentType
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, Http404, HttpResponse
 from django.shortcuts import render, get_object_or_404
 
 from .models import Comment
@@ -8,7 +8,19 @@ from .forms import CommentForm
 
 
 def comment_delete(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    # obj = get_object_or_404(Comment, id=id)
+    # obj = Comment.objects.get(id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+
+    if obj.user != request.user:
+        # messages.success(request, 'You do not have permissions to view this.')
+        # raise Http404
+        response = HttpResponse('You do not have permissions to do this.')
+        response.status_code = 403
+        return response
     if request.method == 'POST':
         parent_obj_url = obj.content_object.get_absolute_url()
         obj.delete()
@@ -21,7 +33,15 @@ def comment_delete(request, id):
 
 
 def comment_thread(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    # obj = Comment.objects.get(id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+
+    if not obj.is_parent:
+        obj = obj.parent
+
     # another way to get data for form below
     # content_object = obj.content_object
     # content_id = obj.content_object.id
